Separate API from main.py so that a demo version can be switched in to avoid needing a RabbitMQ and ISPyB server for testing
